requirements.txt
agno>=1.1.1
mcp>=1.6.0
python-dotenv>=1.1.0
kiteconnect>=5.0.1
starlette>=0.46.1
uvicorn>=0.34.0
rich>=13.7.0
google-adk>=0.1.0
